requirements.txt
Flask==3.0.0
Flask-Session==0.6.0
mysql-connector-python==8.3.0
bcrypt==4.1.2
Werkzeug==3.0.1
